services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.13.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0